An mc-option is now only removed if present in a feedback option

+++ zesje/api/feedback.py
@@ -143,7 +143,8 @@ class Feedback(Resource):
         db.session.commit()
 
         # Delete mc_options associated with this feedback option
-        db.session.delete(fb.mc_option)
-        db.session.commit()
+        if fb.mc_option:
+            db.session.delete(fb.mc_option)
+            db.session.commit()
 
         return dict(status=200, message=f"Feedback option with id {feedback_id} deleted."), 200
